move toAscii to own file
